requirements.txt
requests
psycopg2-binary
pandas
google-api-python-client
google-auth-httplib2
python-dotenv
